+++ app/Resources/doc/changelog.md
@@ -2,7 +2,7 @@
 
 ## 2.0.4.1
 * Refactored dataset and service controllers.
-
+* Minor changes on search bundle.
 
 ## 2.0.4.0
 * Helptext import/export commands.